+++ src/main/java/roomescape/dao/ReservationDAO.java
@@ -1,0 +1,70 @@
+package roomescape.dao;
+
+import java.sql.PreparedStatement;
+import java.util.List;
+
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationCreateResponse;
+import roomescape.dto.TimeCreateResponse;
+
+@Repository
+public class ReservationDAO {
+    private final JdbcTemplate jdbcTemplate;
+
+    public ReservationDAO(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public Long insertWithKeyHolder(Reservation reservation, Long timeId) {
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(
+                    "insert into reservation (name, date, time_id) values (?, ?, ?)",
+                    new String[]{"id"});
+            ps.setString(1, reservation.getName());
+            ps.setString(2, reservation.getDate());
+            ps.setLong(3, timeId);
+            return ps;
+        }, keyHolder);
+
+        return keyHolder.getKey().longValue();
+    }
+
+    public List<ReservationCreateResponse> findAllReservations() {
+        String sql = """
+                SELECT
+                    r.id as reservation_id,
+                    r.name,
+                    r.date,
+                    t.id as time_id,
+                    t.start_at as time_value
+                FROM reservation as r
+                INNER JOIN reservation_time as t
+                  ON r.time_id = t.id
+                """;
+        return jdbcTemplate.query(
+                sql,
+                (resultSet, rowNum) -> {
+                    TimeCreateResponse timeCreateResponse = new TimeCreateResponse(
+                            resultSet.getLong("time_id"),
+                            resultSet.getString("time_value")
+                    );
+                    return new ReservationCreateResponse(
+                            resultSet.getLong("reservation_id"),
+                            resultSet.getString("name"),
+                            resultSet.getString("date"),
+                            timeCreateResponse
+                    );
+                });
+    }
+
+    public int delete(Long id) {
+        String sql = "delete from reservation where id = ?";
+        return jdbcTemplate.update(sql, Long.valueOf(id));
+    }
+}
