--- src/main/java/roomescape/dao/UpdatingDAO.java
@@ -1,58 +1,0 @@
-package roomescape.dao;
-
-import java.sql.PreparedStatement;
-
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Repository;
-
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-
-@Repository
-public class UpdatingDAO {
-    private JdbcTemplate jdbcTemplate;
-
-    public UpdatingDAO(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public int delete(Long id) {
-        String sql = "delete from reservation where id = ?";
-        return jdbcTemplate.update(sql, Long.valueOf(id));
-    }
-
-    public int deleteTime(Long id) {
-        String sql = "delete from reservation_time where id = ?";
-        return jdbcTemplate.update(sql, Long.valueOf(id));
-    }
-
-    public Long insertWithKeyHolder(Reservation reservation, Long timeId) {
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        jdbcTemplate.update(connection -> {
-            PreparedStatement ps = connection.prepareStatement(
-                    "insert into reservation (name, date, time_id) values (?, ?, ?)",
-                    new String[]{"id"});
-            ps.setString(1, reservation.getName());
-            ps.setString(2, reservation.getDate());
-            ps.setLong(3, timeId);
-            return ps;
-        }, keyHolder);
-
-        return keyHolder.getKey().longValue();
-    }
-
-    public Long insertWithKeyHolder(ReservationTime reservationTime) {
-
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-        jdbcTemplate.update(connection -> {
-            PreparedStatement ps = connection.prepareStatement(
-                    "insert into reservation_time (start_at) values (?)",
-                    new String[]{"id"});
-            ps.setString(1, reservationTime.getStartAt());
-            return ps;
-        }, keyHolder);
-        return keyHolder.getKey().longValue();
-    }
-}
